fix(storage): cast SUM() results to BIGINT in finance repo queries
PostgreSQL SUM() on BIGINT columns returns NUMERIC, causing sqlx type
mismatch when deserializing into i64. Add ::BIGINT casts to all 8
aggregate queries across 5 finance repos.

+++ crates/storage/src/repos/finance_budget_repo.rs
@@ -168,7 +168,7 @@ impl FinanceBudgetRepo {
                 b.alert_threshold,
                 b.created_at,
                 b.updated_at,
-                COALESCE(SUM(ft.amount), 0) AS spent
+                COALESCE(SUM(ft.amount), 0)::BIGINT AS spent
             FROM finance_budgets b
             LEFT JOIN finance_transactions ft ON
                 ft.tx_type = 'expense'
@@ -216,7 +216,7 @@ impl FinanceBudgetRepo {
                 b.alert_threshold,
                 b.created_at,
                 b.updated_at,
-                COALESCE(SUM(ft.amount), 0) AS spent
+                COALESCE(SUM(ft.amount), 0)::BIGINT AS spent
             FROM finance_budgets b
             LEFT JOIN finance_transactions ft ON
                 ft.tx_type = 'expense'
